+++ crates/honk-engine/src/interaction.rs
@@ -33,9 +33,9 @@ impl Default for Pointer {
 /// Hover-movement (px) that registers one "pat".
 pub const SWEEP_PER_PAT: f32 = 28.0;
 /// With no new pat within this many seconds, the streak resets to zero.
-pub const STREAK_TIMEOUT: f32 = 1.2;
+pub const STREAK_TIMEOUT: f64 = 1.2;
 /// Each registered pat keeps the goose calm for this long (seconds).
-pub const CALM_DURATION: f32 = 4.0;
+pub const CALM_DURATION: f64 = 4.0;
 
 /// Detects pats from successive hovering cursor positions and tracks the happy streak.
 #[derive(Debug, Clone)]
@@ -43,8 +43,8 @@ pub struct PatTracker {
     streak: u32,
     sweep_accum: f32,
     last_pos: Option<Vec2>,
-    last_pat_time: f32,
-    calm_until: f32,
+    last_pat_time: f64,
+    calm_until: f64,
 }
 
 impl Default for PatTracker {
@@ -53,8 +53,8 @@ impl Default for PatTracker {
             streak: 0,
             sweep_accum: 0.0,
             last_pos: None,
-            last_pat_time: f32::NEG_INFINITY,
-            calm_until: f32::NEG_INFINITY,
+            last_pat_time: f64::NEG_INFINITY,
+            calm_until: f64::NEG_INFINITY,
         }
     }
 }
@@ -71,13 +71,13 @@ impl PatTracker {
     }
 
     /// Whether the goose is in its post-pat calm window at `now`.
-    pub fn is_calm(&self, now: f32) -> bool {
+    pub fn is_calm(&self, now: f64) -> bool {
         now < self.calm_until
     }
 
     /// Feed one frame of pointer state. `hovering` is whether the cursor is over the goose.
     /// Returns how many pats were registered this frame (0 normally, ≥1 on a long sweep).
-    pub fn update(&mut self, hovering: bool, pos: Vec2, now: f32) -> u32 {
+    pub fn update(&mut self, hovering: bool, pos: Vec2, now: f64) -> u32 {
         // Lapse the streak after a quiet spell (the in-progress sweep keeps accumulating —
         // it only resets when the cursor actually leaves the goose, below).
         if now - self.last_pat_time > STREAK_TIMEOUT {
@@ -112,7 +112,7 @@ mod tests {
     use super::*;
 
     /// Drag the cursor a total of `dist` px across the goose in one straight sweep.
-    fn sweep(t: &mut PatTracker, start: Vec2, dist: f32, now: f32) -> u32 {
+    fn sweep(t: &mut PatTracker, start: Vec2, dist: f32, now: f64) -> u32 {
         // Prime the baseline position (first hovering frame moves nothing).
         t.update(true, start, now);
         t.update(true, start + Vec2::new(dist, 0.0), now)
